chore: moving some mutations where they are used
