+++ app/graphql/operations/remote/mutations.ts
@@ -1,26 +1,9 @@
 import { gql } from "@apollo/client"
 
 export default gql`
-  mutation captchaCreateChallenge {
-    captchaCreateChallenge {
-      errors {
-        __typename
-        message
-      }
-      result {
-        __typename
-        id
-        challengeCode
-        newCaptcha
-        failbackMode
-      }
-    }
-  }
-
   mutation userContactUpdateAlias($input: UserContactUpdateAliasInput!) {
     userContactUpdateAlias(input: $input) {
       errors {
-        __typename
         message
       }
     }
@@ -29,30 +12,9 @@ export default gql`
   mutation deviceNotificationTokenCreate($input: DeviceNotificationTokenCreateInput!) {
     deviceNotificationTokenCreate(input: $input) {
       errors {
-        __typename
         message
       }
       success
-    }
-  }
-
-  mutation intraLedgerPaymentSend($input: IntraLedgerPaymentSendInput!) {
-    intraLedgerPaymentSend(input: $input) {
-      errors {
-        __typename
-        message
-      }
-      status
-    }
-  }
-
-  mutation intraLedgerUsdPaymentSend($input: IntraLedgerUsdPaymentSendInput!) {
-    intraLedgerUsdPaymentSend(input: $input) {
-      errors {
-        __typename
-        message
-      }
-      status
     }
   }
 
@@ -61,7 +23,6 @@ export default gql`
   ) {
     userQuizQuestionUpdateCompleted(input: $input) {
       errors {
-        __typename
         message
       }
       userQuizQuestion {
@@ -74,34 +35,9 @@ export default gql`
     }
   }
 
-  mutation userUpdateUsername($input: UserUpdateUsernameInput!) {
-    userUpdateUsername(input: $input) {
-      errors {
-        __typename
-        message
-      }
-      user {
-        __typename
-        id
-        username
-      }
-    }
-  }
-
-  mutation captchaRequestAuthCode($input: CaptchaRequestAuthCodeInput!) {
-    captchaRequestAuthCode(input: $input) {
-      errors {
-        __typename
-        message
-      }
-      success
-    }
-  }
-
   mutation lnNoAmountInvoiceFeeProbe($input: LnNoAmountInvoiceFeeProbeInput!) {
     lnNoAmountInvoiceFeeProbe(input: $input) {
       errors {
-        __typename
         message
       }
       amount
@@ -111,7 +47,6 @@ export default gql`
   mutation lnInvoiceFeeProbe($input: LnInvoiceFeeProbeInput!) {
     lnInvoiceFeeProbe(input: $input) {
       errors {
-        __typename
         message
       }
       amount
@@ -121,7 +56,6 @@ export default gql`
   mutation lnUsdInvoiceFeeProbe($input: LnUsdInvoiceFeeProbeInput!) {
     lnUsdInvoiceFeeProbe(input: $input) {
       errors {
-        __typename
         message
       }
       amount
@@ -131,74 +65,18 @@ export default gql`
   mutation lnNoAmountUsdInvoiceFeeProbe($input: LnNoAmountUsdInvoiceFeeProbeInput!) {
     lnNoAmountUsdInvoiceFeeProbe(input: $input) {
       errors {
-        __typename
         message
       }
       amount
     }
   }
 
-  mutation userUpdateUsername($input: UserUpdateUsernameInput!) {
-    userUpdateUsername(input: $input) {
-      errors {
-        __typename
-        message
-      }
-      user {
-        __typename
-        id
-        username
-      }
-    }
-  }
-
   mutation userLogin($input: UserLoginInput!) {
     userLogin(input: $input) {
       errors {
-        __typename
         message
       }
       authToken
-    }
-  }
-
-  mutation lnNoAmountInvoicePaymentSend($input: LnNoAmountInvoicePaymentInput!) {
-    lnNoAmountInvoicePaymentSend(input: $input) {
-      errors {
-        __typename
-        message
-      }
-      status
-    }
-  }
-
-  mutation lnInvoicePaymentSend($input: LnInvoicePaymentInput!) {
-    lnInvoicePaymentSend(input: $input) {
-      errors {
-        __typename
-        message
-      }
-      status
-    }
-  }
-
-  mutation lnNoAmountUsdInvoicePaymentSend($input: LnNoAmountUsdInvoicePaymentInput!) {
-    lnNoAmountUsdInvoicePaymentSend(input: $input) {
-      errors {
-        __typename
-        message
-      }
-      status
-    }
-  }
-
-  mutation onChainPaymentSend($input: OnChainPaymentSendInput!) {
-    onChainPaymentSend(input: $input) {
-      errors {
-        __typename
-        message
-      }
-      status
     }
   }
 `
