+++ app/screens/galoy-address-screen/set-address-modal.tsx
@@ -13,6 +13,7 @@ import useMainQuery from "@app/hooks/use-main-query"
 import { useAppConfig } from "@app/hooks/use-app-config"
 import { useUserUpdateUsernameMutation } from "@app/graphql/generated"
 import { bankName } from "@app/config"
+import { gql } from "@apollo/client"
 
 const styles = EStyleSheet.create({
   centeredView: {
@@ -115,6 +116,20 @@ type SetAddressModalProps = {
   modalVisible: boolean
   toggleModal?: () => void
 }
+
+gql`
+  mutation userUpdateUsername($input: UserUpdateUsernameInput!) {
+    userUpdateUsername(input: $input) {
+      errors {
+        message
+      }
+      user {
+        id
+        username
+      }
+    }
+  }
+`
 
 export const SetAddressModal = ({ modalVisible, toggleModal }: SetAddressModalProps) => {
   const { LL } = useI18nContext()
