+++ app/screens/send-bitcoin-screen/send-bitcoin-confirmation-screen.tsx
@@ -33,6 +33,7 @@ import { FakeCurrencyInput } from "react-native-currency-input"
 import { testProps } from "../../../utils/testProps"
 import { Status } from "./send-bitcoin.types"
 import useFee from "./use-fee"
+import { gql } from "@apollo/client"
 
 const styles = StyleSheet.create({
   scrollView: {
@@ -181,6 +182,62 @@ const styles = StyleSheet.create({
     alignItems: "center",
   },
 })
+
+gql`
+  mutation intraLedgerPaymentSend($input: IntraLedgerPaymentSendInput!) {
+    intraLedgerPaymentSend(input: $input) {
+      errors {
+        message
+      }
+      status
+    }
+  }
+
+  mutation intraLedgerUsdPaymentSend($input: IntraLedgerUsdPaymentSendInput!) {
+    intraLedgerUsdPaymentSend(input: $input) {
+      errors {
+        message
+      }
+      status
+    }
+  }
+
+  mutation lnNoAmountInvoicePaymentSend($input: LnNoAmountInvoicePaymentInput!) {
+    lnNoAmountInvoicePaymentSend(input: $input) {
+      errors {
+        message
+      }
+      status
+    }
+  }
+
+  mutation lnInvoicePaymentSend($input: LnInvoicePaymentInput!) {
+    lnInvoicePaymentSend(input: $input) {
+      errors {
+        message
+      }
+      status
+    }
+  }
+
+  mutation lnNoAmountUsdInvoicePaymentSend($input: LnNoAmountUsdInvoicePaymentInput!) {
+    lnNoAmountUsdInvoicePaymentSend(input: $input) {
+      errors {
+        message
+      }
+      status
+    }
+  }
+
+  mutation onChainPaymentSend($input: OnChainPaymentSendInput!) {
+    onChainPaymentSend(input: $input) {
+      errors {
+        message
+      }
+      status
+    }
+  }
+`
 
 const SendBitcoinConfirmationScreen = ({
   navigation,
